+++ core/src/main/java/org/extism/sdk/chicory/core/HttpConfig.java
@@ -2,7 +2,6 @@ package org.extism.sdk.chicory.core;
 
 import org.extism.sdk.chicory.http.HttpClientAdapter;
 import org.extism.sdk.chicory.http.HttpJsonCodec;
-import org.extism.sdk.chicory.http.HttpUrlConnectionClientAdapter;
 import org.extism.sdk.chicory.http.JdkHttpClientAdapter;
 
 import java.util.Objects;
@@ -13,21 +12,21 @@ public class HttpConfig {
      * Use {@link JdkHttpClientAdapter} for the HTTP client adapter.
      * Recommended on recent Java versions.
      */
-    public static HttpConfig defaultConfig() {
-        return HttpConfig.builder()
-                .withClientAdapter(JdkHttpClientAdapter::new)
-                .withJsonCodec(JacksonJsonCodec::new).build();
-    }
-
-    /**
-     * Use {@link HttpUrlConnectionClientAdapter} for the HTTP client adapter.
-     * Recommended for Android.
-     */
-    public static HttpConfig urlConnectionConfig() {
-        return HttpConfig.builder()
-                .withClientAdapter(HttpUrlConnectionClientAdapter::new)
-                .withJsonCodec(JakartaJsonCodec::new).build();
-    }
+//    public static HttpConfig defaultConfig() {
+//        return HttpConfig.builder()
+//                .withClientAdapter(JdkHttpClientAdapter::new)
+//                .withJsonCodec(JacksonJsonCodec::new).build();
+//    }
+//
+//    /**
+//     * Use {@link HttpUrlConnectionClientAdapter} for the HTTP client adapter.
+//     * Recommended for Android.
+//     */
+//    public static HttpConfig urlConnectionConfig() {
+//        return HttpConfig.builder()
+//                .withClientAdapter(HttpUrlConnectionClientAdapter::new)
+//                .withJsonCodec(JakartaJsonCodec::new).build();
+//    }
 
     public static Builder builder() {
         return new Builder();
