+++ http/integration-tests/pom.xml
@@ -9,7 +9,7 @@
         <version>999-SNAPSHOT</version>
     </parent>
 
-    <artifactId>http-jakarta-json</artifactId>
+    <artifactId>http-integration-tests</artifactId>
 
     <properties>
         <maven.compiler.source>22</maven.compiler.source>
@@ -17,17 +17,26 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
-
     <dependencies>
         <dependency>
             <groupId>org.extism.sdk</groupId>
-            <artifactId>http-api</artifactId>
-            <version>${project.version}</version>
+            <artifactId>chicory-sdk-core</artifactId>
+            <version>999-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>jakarta.json</groupId>
-            <artifactId>jakarta.json-api</artifactId>
-            <version>${jakarta.json-api.version}</version>
+            <groupId>org.extism.sdk</groupId>
+            <artifactId>http-api</artifactId>
+            <version>999-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.extism.sdk</groupId>
+            <artifactId>http-jackson</artifactId>
+            <version>999-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.extism.sdk</groupId>
+            <artifactId>http-jakarta-json</artifactId>
+            <version>999-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.eclipse.parsson</groupId>
@@ -35,6 +44,13 @@
             <version>${jakarta.json.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>${junit.version}</version>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 
 </project>
