+++ http/pom.xml
@@ -15,6 +15,7 @@
         <module>jackson</module>
         <module>jakarta-json</module>
         <module>api</module>
+        <module>integration-tests</module>
     </modules>
 
     <properties>
